Massive work in progress.
Attempting to get animations to a workable state.

Most of these changes will be backed out later.

+++ ProjectBarnabus/src/GameEngine/FreeCamera.cpp
@@ -73,7 +73,7 @@ void FreeCamera::Update(double deltaTime)
 	glfwGetCursorPos(BarnabusGameEngine::Get().GetWindow(), &cursorX, &cursorY);
 }
 
-void FreeCamera::Move(CameraMovement direction, float dist)
+void FreeCamera::Move(CameraMovement direction, double dist)
 {
 	switch (direction)
 	{
